repeat sent message, use Cookies, update logger.

+++ src/types.rs
@@ -14,12 +14,12 @@ pub struct ChatView {
 #[derive(Template)]
 #[template(path = "sse.html")]
 pub struct SseView {
-    pub from_user_id: usize,
+    pub from_user_id: u16,
     pub data: String,
 }
 
 #[derive(Deserialize)]
 pub struct SendMsgReq {
-    pub target_id: usize,
+    pub target_id: u16,
     pub message: String,
 }
